cpu-llm/requirements.txt
llama-cpp-python==0.2.37
huggingface_hub